Cast macosx_deployment_target to string
Related to #54

Big Sur has a deployment target of `11` as opposed to `10.16` or `11.0`. Therefore `sysconfig.get_config_var()` converts it to an int. However, setting `os.environ[]` requires a string so it's raising an exception.

There seems to be a discussion and some disagreement within Python (https://bugs.python.org/issue42504) on if this should be forced to a string or not, but the conversation died at the beginning of December.

Even if they decide to make it a string again it'll take awhile before it's available so I think it may be best to add the casting.